requirements.txt
python-dotenv==0.19.2
dict2xml==1.7.0
requests==2.27.1